+++ profiling/src/profile/internal/mod.rs
@@ -1,6 +1,7 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2023-Present Datadog, Inc.
 
+mod endpoints;
 mod function;
 mod label;
 mod location;
@@ -11,6 +12,7 @@ mod string;
 mod upscaling;
 mod value_type;
 
+pub use endpoints::*;
 pub use function::*;
 pub use label::*;
 pub use location::*;
